Update SpecRankings to read/write from S3

+++ lorgs/clients/wcl/models/character_ranking.py
@@ -35,7 +35,7 @@ class CharacterRanking(BaseModel):
     amount: float
     """total DPS/HPS."""
 
-    duration: float
+    duration: int
     """Fight duration in Milliseconds."""
 
     report: CharacterRankingReportFightData
